Refactor versions
